requirements.txt
beautifulsoup4==4.9.3
jupyter==1.0.0
jupyter-client==6.1.10
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
pandas==1.2.0
Pillow==7.0.0
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
tweepy==3.10.0


